add note on builder pnl calc limitations
Former-commit-id: 86108d828f13acba04625f6b54b51cbac2bca373

+++ crates/brontes-inspect/src/mev_inspectors/shared_utils.rs
@@ -499,7 +499,7 @@ impl<DB: LibmdbxReader> SharedInspectorUtils<'_, DB> {
             .db
             .get_protocol_details(address)
             .ok()
-            .map(|protocol| protocol.protocol.short().to_string());
+            .map(|protocol| protocol.protocol.to_string());
 
         protocol_name.or_else(|| {
             self.db
